Add option to avoid parsing API response
This is useful for clients that do not need to parse the results
right away and want to save the results in their raw format, or
if a client wants to handle error codes directly rather than
raising an exception.

It also provides an option for accessing new fields that have been
added to the API response but haven't been added to the parser for
this gem.